+++ ModReleasePackage/CHANGELOG.md
@@ -1,3 +1,9 @@
-# 1.0.0
+# 1.1.0
+* Make some members internal
+* Add a bunch of missing members. Probably should have not uploaded this yet. Oops.
+* Added the option (on by default) to disable the Low Pass Filter after dying when the death state is known to be a manually defined Void death.
+* Added the option (on by default) to register all known vanilla void enemies for scripted immunity to fog. This prevents weird edge cases with some void spawns taking damage in their own atmosphere.
+* Added the option to globally configure the mechanics of player-created void black holes. The options are the same as the original Reaver mod, where instakills to monsters can be toggled, instakills to bosses can be toggled, and friendly fire can be toggled. The fallback damage (when an instakill isn't possible) is also a configurable value.
 
+# 1.0.0
 * Initial release of the mod.